+++ example/queries.dart
@@ -7,7 +7,7 @@ main() async {
   await db.open();
   print('connection open');
   coll = db.collection("simple_data");
-  await coll.remove();
+  await coll.remove({});
   print('Packing data to insert into collection by Bson...');
   for (var n = 0; n < 1000; n++) {
     await coll.insert({"my_field": n, "str_field": "str_$n"});