+++ example/russian.dart
@@ -4,7 +4,7 @@ main() async {
   Db db = new Db("mongodb://127.0.0.1/mongo_dart-test");
   await db.open();
   DbCollection collection = db.collection('test-utf8');
-  await collection.remove();
+  await collection.remove({});
   await collection.insert({
     'Имя': 'Вадим',
     'Фамилия': 'Цушко',