+++ test/replica_tests.dart
@@ -18,7 +18,7 @@ Future testCollectionInfoCursor() {
   DbCollection newColl;
   return db.open(writeConcern: WriteConcern.JOURNALED).then((c) {
     newColl = db.collection("new_collecion");
-    return newColl.remove();
+    return newColl.remove({});
   }).then((v) {
     return newColl.insertAll([
       {"a": 1}